Moved methods from Utility into ArrowheadProperties
